fix clang headers to support HIP c kernels and to support OpenMP .  This depricates the autoheader __clang_openmp_runtime_wrapper.h .  Both nvptx and amdgcn will use the autoheader opemp_wrappers/__clang_openmp_device-functions.h
