+++ clang/lib/Driver/ToolChains/AMDGPUOpenMP.cpp
@@ -170,10 +170,6 @@ const char *AMDGCN::OpenMPLinker::constructOmpExtraCmds(
          "hip.bc", "ockl.bc",
          Args.MakeArgString("libbc-hostrpc-amdgcn.a"),
          std::string(WaveFrontSizeBC)});
-
-    if (!Args.hasArg(options::OPT_nostdlibxx) &&
-        !Args.hasArg(options::OPT_nostdlib))
-      BCLibs.append({Args.MakeArgString("libm-amdgcn-" + SubArchName + ".bc")});
   }
 
   for (auto Lib : BCLibs)
@@ -562,6 +558,11 @@ void AMDGPUOpenMPToolChain::AddClangSystemIncludeArgs(const ArgList &DriverArgs,
   CC1Args.push_back(DriverArgs.MakeArgString(D.Dir + "/../../include"));
 
   HostTC.AddClangSystemIncludeArgs(DriverArgs, CC1Args);
+
+  CC1Args.push_back("-internal-isystem");
+  SmallString<128> P(HostTC.getDriver().ResourceDir);
+  llvm::sys::path::append(P, "include/cuda_wrappers");
+  CC1Args.push_back(DriverArgs.MakeArgString(P));
 }
 
 /// Convert path list to Fortran frontend argument
