+++ clang/lib/Headers/CMakeLists.txt
@@ -48,8 +48,9 @@ set(files
   __clang_cuda_runtime_wrapper.h
   __clang_hip_libdevice_declares.h
   __clang_hip_math.h
+  __clang_hip_cmath.h
   __clang_hip_runtime_wrapper.h
-  __clang_openmp_runtime_wrapper.h
+  __clang_hip_complex_builtins.h
   cetintrin.h
   cet.h
   cldemoteintrin.h
