+++ clang/lib/Headers/openmp_wrappers/cmath
@@ -72,4 +72,13 @@ __DEVICE__ float tgamma(float __x) { return ::tgammaf(__x); }
 
 #pragma omp end declare variant
 
+#pragma omp begin declare variant match(                                       \
+    device = {arch(amdgcn)}, implementation = {extension(match_any)})
+#define __HIP__
+#define __OPENMP_AMDGCN__
+#include <__clang_hip_cmath.h>
+#undef __OPENMP_AMDGCN__
+#undef __HIP__
+#pragma omp end declare variant
+
 #endif
